Move the keyboard-shortcuts dialog into its own component
130 lines of static key-hint table leave App.vue's template. The dialog
had no coverage of any kind, and the extraction is exactly the sort that
fails silently, so it gains an e2e spec pinning what can break: F1
reaching the handler, and the table rendering rows.

App.vue is now 771 lines, under the 800 the refactor plan asks for, down
from 2455.
