+++ frontend/src/App.vue
@@ -31,7 +31,6 @@ import {
   UNASSIGNED_PICTURES_ID,
   useViewStore,
 } from "./stores/useViewStore";
-import { redoKeyHint, undoKeyHint } from "./utils/shortcutHints";
 import { useAppConfig } from "./composables/useAppConfig";
 import { useAppNavigation } from "./composables/useAppNavigation";
 import { useGlobalKeydown } from "./composables/useGlobalKeydown";
@@ -52,6 +51,7 @@ import ReviewSessionsOverlay from "./components/views/ReviewSessionsOverlay.vue"
 import StatsSidebar from "./components/panels/StatsSidebar.vue";
 import ThumbnailUpgradeBanner from "./components/panels/ThumbnailUpgradeBanner.vue";
 import NoticeHost from "./components/widgets/NoticeHost.vue";
+import ShortcutsDialog from "./components/widgets/ShortcutsDialog.vue";
 import { useFloatingBottomInset } from "./composables/useBottomAnchor";
 import { toPx } from "./utils/floatingBottom.js";
 const BACKEND_URL = API_BASE_URL;
@@ -76,8 +76,6 @@ const operationStore = useOperationStore();
 const viewStore = useViewStore();
 // Keycap labels for the shortcuts dialog. The binding accepts Ctrl and Meta
 // everywhere; only the hint is platform-specific.
-const undoKeyHintKeys = undoKeyHint();
-const redoKeyHintKeys = redoKeyHint();
 
 // --- Router ---
 const route = useRoute();
@@ -142,11 +140,8 @@ const {
   refreshSidebarPicturesDebounced,
 } = useSidebarRefresh({ sidebarRef });
 
-const {
-  updateIsMobile,
-  updateMaxColumns,
-  closeSidebarIfMobile,
-} = useViewportLayout({ mainAreaRef });
+const { updateIsMobile, updateMaxColumns, closeSidebarIfMobile } =
+  useViewportLayout({ mainAreaRef });
 
 // The live-updates channel. App.vue owns its lifecycle - it connects on
 // mount and disconnects on unmount - but the socket, its filter handshake and
@@ -775,136 +770,7 @@ defineExpose({
     >
       <v-icon size="20">mdi-keyboard</v-icon><span>F1</span>
     </button>
-    <v-dialog v-model="shortcutsDialogOpen" max-width="480">
-      <v-card class="shortcuts-dialog">
-        <v-card-title class="shortcuts-dialog-title"
-          >Keyboard shortcuts</v-card-title
-        >
-        <v-card-text class="shortcuts-dialog-body">
-          <table class="shortcuts-table">
-            <tbody>
-              <tr>
-                <td colspan="2" class="shortcuts-section">Grid view</td>
-              </tr>
-              <tr>
-                <td><kbd>F</kbd></td>
-                <td>Open search</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>1</kbd> – <kbd>5</kbd></td>
-                <td>Set star rating on hovered / selected image(s)</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>T</kbd></td>
-                <td>Tag selected images</td>
-              </tr>
-              <tr>
-                <td><kbd>Ctrl</kbd>+<kbd>A</kbd></td>
-                <td>Select all images</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td>
-                  <template v-for="(key, i) in undoKeyHintKeys" :key="key"
-                    ><span v-if="i > 0">+</span><kbd>{{ key }}</kbd></template
-                  >
-                </td>
-                <td>Undo the last change</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td>
-                  <template v-for="(key, i) in redoKeyHintKeys" :key="key"
-                    ><span v-if="i > 0">+</span><kbd>{{ key }}</kbd></template
-                  >
-                </td>
-                <td>Redo the change you just undid</td>
-              </tr>
-              <tr>
-                <td><kbd>G</kbd></td>
-                <td>Focus first visible image (start keyboard navigation)</td>
-              </tr>
-              <tr>
-                <td><kbd>←</kbd> <kbd>→</kbd> <kbd>↑</kbd> <kbd>↓</kbd></td>
-                <td>Move cursor and select image</td>
-              </tr>
-              <tr>
-                <td><kbd>Shift</kbd>+<kbd>Arrow</kbd></td>
-                <td>Extend selection</td>
-              </tr>
-              <tr>
-                <td><kbd>Ctrl</kbd>+<kbd>Arrow</kbd></td>
-                <td>Move cursor without changing selection</td>
-              </tr>
-              <tr>
-                <td><kbd>Space</kbd></td>
-                <td>Toggle selection of cursor image</td>
-              </tr>
-              <tr>
-                <td><kbd>Enter</kbd></td>
-                <td>Open cursor image</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>Delete</kbd></td>
-                <td>Delete selected images</td>
-              </tr>
-              <tr>
-                <td><kbd>Esc</kbd></td>
-                <td>Clear selection</td>
-              </tr>
-              <tr>
-                <td><kbd>S</kbd></td>
-                <td>Open selection menu</td>
-              </tr>
-              <tr>
-                <td><kbd>Home</kbd> / <kbd>End</kbd></td>
-                <td>Jump to first / last image</td>
-              </tr>
-              <tr>
-                <td><kbd>Page Up</kbd> / <kbd>Page Down</kbd></td>
-                <td>Scroll image grid</td>
-              </tr>
-              <tr>
-                <td colspan="2" class="shortcuts-section">Image overlay</td>
-              </tr>
-              <tr>
-                <td><kbd>←</kbd> <kbd>→</kbd></td>
-                <td>Previous / next image</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>1</kbd> – <kbd>5</kbd></td>
-                <td>Set star rating</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>T</kbd></td>
-                <td>Add tag</td>
-              </tr>
-              <tr>
-                <td><kbd>Z</kbd></td>
-                <td>Toggle zoom</td>
-              </tr>
-              <tr>
-                <td><kbd>I</kbd></td>
-                <td>Toggle info panel</td>
-              </tr>
-              <tr>
-                <td><kbd>Esc</kbd></td>
-                <td>Close overlay</td>
-              </tr>
-              <tr>
-                <td colspan="2" class="shortcuts-section">General</td>
-              </tr>
-              <tr :class="{ 'shortcut-disabled': isReadOnly }">
-                <td><kbd>F2</kbd></td>
-                <td>Edit selected character or picture set</td>
-              </tr>
-              <tr>
-                <td><kbd>?</kbd> / <kbd>F1</kbd></td>
-                <td>Show / hide this dialog</td>
-              </tr>
-            </tbody>
-          </table>
-        </v-card-text>
-      </v-card>
-    </v-dialog>
+    <ShortcutsDialog v-model="shortcutsDialogOpen" />
   </v-app>
 </template>
 <style src="./App.css"></style>
